Refactor validation error handling
