+++ src/helpers/baseHelper.js
@@ -1,5 +1,15 @@
 import bcrypt from 'bcryptjs';
+import { validationResult } from 'express-validator';
 
+export const handleErrors = (req, res, next) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({
+      error: errors.array().map((i) => i.msg),
+    });
+  }
+  return next();
+};
 export const hashPassword = (password) => {
   const salt = bcrypt.genSaltSync(10);
   const hash = bcrypt.hashSync(password, salt);
